AG - Moved customer manage billing tests to its own class and integrated Lj's customer page with my own
